Move some interactive sql shell command to gcp/cloudsql/interactive.ts
