+++ src/gcp/cloudsql/interactive.ts
@@ -1,0 +1,55 @@
+import * as pg from "pg";
+import * as ora from "ora";
+import chalk from "chalk";
+import { logger } from "../../logger";
+import { confirm } from "../../prompt";
+
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const Table = require("cli-table");
+
+// Not comprehensive list, used for best offer prompting.
+const destructiveSqlKeywords = ["DROP", "DELETE"];
+
+function checkIsDestructiveSql(query: string): boolean {
+  const upperCaseQuery = query.toUpperCase();
+  return destructiveSqlKeywords.some((keyword) => upperCaseQuery.includes(keyword.toUpperCase()));
+}
+
+export async function confirmDangerousQuery(query: string): Promise<boolean> {
+  if (checkIsDestructiveSql(query)) {
+    return await confirm({
+      message: chalk.yellow("This query may be destructive. Are you sure you want to proceed?"),
+      default: false,
+    });
+  }
+  return true;
+}
+
+// Pretty query execution display such as spinner and actual returned content for `SELECT` query.
+export async function interactiveExecuteQuery(query: string, conn: pg.PoolClient) {
+  const spinner = ora("Executing query...").start();
+  try {
+    const results = await conn.query(query);
+    spinner.succeed(chalk.green("Query executed successfully"));
+
+    if (Array.isArray(results.rows) && results.rows.length > 0) {
+      const table: any[] = new Table({
+        head: Object.keys(results.rows[0]).map((key) => chalk.cyan(key)),
+        style: { head: [], border: [] },
+      });
+
+      for (const row of results.rows) {
+        table.push(Object.values(row) as any);
+      }
+
+      logger.info(table.toString());
+    } else {
+      // If nothing is returned and the query was select, let the user know there was no results.
+      if (query.toUpperCase().includes("SELECT")) {
+        logger.info(chalk.yellow("No results returned"));
+      }
+    }
+  } catch (err) {
+    spinner.fail(chalk.red(`Failed executing query: ${err}`));
+  }
+}
